requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.18.1
jedi==0.19.2
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
mistralai==1.9.3
nest-asyncio==1.6.0
openai==1.98.0
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.3
pyzmq==27.0.0
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.23.0
